+++ packages/components/src/layout/TotalLayout/TotalLayoutStepContainer/TotalLayoutStepContainer.constants.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+
+export const TOTAL_LAYOUT_STEP_CONTAINER_PROP_TYPES = {
+  stepName: PropTypes.string,
+};
+export const TOTAL_LAYOUT_STEP_CONTAINER_DEFAULT_PROPS = {
+  stepName: '',
+};