+++ packages/components/src/layout/TotalLayout/index.js
@@ -1,0 +1,4 @@
+export * from './TotalLayout';
+export * from './TotalLayoutHeader/TotalLayoutHeader';
+export * from './TotalLayoutStepContainer/TotalLayoutStepContainer';
+export * from './TotalLayout.constants';