+++ packages/notifications/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @bubbles-ui/notifications
 
+## 1.2.52
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.52
+  - @bubbles-ui/icons@1.2.52
+
+## 1.2.51
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.51
+  - @bubbles-ui/icons@1.2.51
+
+## 1.2.50
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.50
+  - @bubbles-ui/icons@1.2.50
+
+## 1.2.49
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.49
+  - @bubbles-ui/icons@1.2.49
+
+## 1.2.48
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.48
+  - @bubbles-ui/icons@1.2.48
+
+## 1.2.47
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.47
+  - @bubbles-ui/icons@1.2.47
+
 ## 1.2.46
 
 ### Patch Changes